pyproject.toml
[project]
authors = [
  {name = "yuvalrot", email = "yuvalrot@telhai.ac.il"},
]
dependencies = [
  "pydantic>=2.12.5",
  "pycountry>=24.6.1",
  "yarl>=1.22.0",
  "miniaudio>=1.61",
  "numpy>=2.3.5",
  "requests>=2.32.5",
  "requests-cache>=1.2.1",
  "requests-ratelimiter>=0.7.0",
  "tqdm>=4.67.1",
]
description = "An intuitive, type-safe Python wrapper for the Xeno-Canto API"
license = {text = "MIT"}
name = "xc-api-py"
readme = "README.md"
requires-python = ">=3.11"
version = "0.1.18"

[tool.pdm]
distribution = true
src-paths = ["src"]

[dependency-groups]
dev = [
  "pytest>=9.0.1",
  "python-dotenv>=1.2.1",
  "rich>=14.2.0",
  "pandas>=2.3.3",
  "sounddevice>=0.5.3",
  "requests-mock>=1.12.1",
]

[project.urls]
"Bug Reports" = "https://github.com/yuvalrot479/xc-api-py/issues"
Homepage = "https://github.com/yuvalrot479/xc-api-py"
Source = "https://github.com/yuvalrot479/xc-api-py"

[tool.ruff]
# Global settings
indent-width = 2
line-length = 120
target-version = "py311"

[tool.ruff.format]
indent-style = "space"
quote-style = "single"
